# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "BE JS Challenge",
      "version": "1.0.11"
    },
...
    "servers": [
      {
        "url": "https://localhost:5001"
      }
    ],
...
    "components": {
      "schemas": {
...
        "transactionModel": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "source_account_id": {
              "type": "integer"
            },
            "destination_account_id": {
              "type": "integer"
            },
            "amount": {
              "type": "integer"
            }
          }
        }
      }
    },
    "paths": {
...
      "/api/v1/transactions": {
        "post": {
          "tags": [
            "transactions"
          ],
          "summary": "Create a transaction",
          "description": "Crate a new transaction with source account and destination account",
          "operationId": "createTransaction",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "amount": {
                      "type": "integer"
                    },
                    "sender_id": {
                      "type": "integer"
                    },
                    "receiver_id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "source_account_id": {
                            "type": "integer"
                          },
                          "destination_account_id": {
                            "type": "integer"
                          },
                          "amount": {
                            "type": "integer"
                          }
                        }
                      },
                      "sender": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      },
                      "receiver": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "bank_name": {
                            "type": "string"
                          },
                          "bank_account_number": {
                            "type": "string"
                          },
                          "balance": {
                            "type": "integer"
                          },
                          "user_id": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "transactions"
          ],
          "summary": "Get all transactions",
          "description": "Return a list of all valid transactions",
          "operationId": "getTransactions",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/transactionModel"
                        }
                      }
                    }
                  }
                }
              }
            },
            "405": {
              "description": "Error",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/transactions/{transaction}": {
        "get": {
          "tags": [
            "transactions"
          ],
          "summary": "Find transaction by ID",
          "description": "Returns a valid transaction",
          "operationId": "getTransactionById",
          "parameters": [
            {
              "name": "transaction",
              "in": "path",
              "description": "ID of transaction to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "data": {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "source_account_id": {
                                "type": "integer"
                              },
                              "destination_account_id": {
                                "type": "integer"
                              },
                              "amount": {
                                "type": "integer"
                              }
                            }
                          },
                          "source": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "bank_name": {
                                "type": "string"
                              },
                              "bank_account_number": {
                                "type": "string"
                              },
                              "balance": {
                                "type": "integer"
                              },
                              "user_id": {
                                "type": "integer"
                              }
                            }
                          },
                          "destination": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "bank_name": {
                                "type": "string"
                              },
                              "bank_account_number": {
                                "type": "string"
                              },
                              "balance": {
                                "type": "integer"
                              },
                              "user_id": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }